requirements.txt
# --------- pytorch --------- #
torch==1.8.1
torchvision==0.9.1
pytorch-lightning>=1.2.7

# --------- hydra --------- #
hydra-core==1.1.0.dev5
hydra-colorlog==1.1.0.dev1
hydra-optuna-sweeper==1.1.0.dev1
# hydra-ray-launcher==0.1.2
# hydra-submitit-launcher==1.1.0

# --------- loggers --------- #
wandb>=0.10.26
# neptune-client
# comet-ml
# mlflow

# --------- others --------- #
albumentations>=0.5.2
rich>=10.1.0
python-dotenv >= 0.17.0
scikit-learn>= 0.24.0
pandas>=1.2.4
matplotlib>=3.4.1
seaborn>=0.11.1
pytest>=6.2.3
# --------- personal packages --------- #
cv-models>=0.1.0